requirements.txt
appnope==0.1.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.0
Django==2.1.4
django-filter==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.2.0
get==2019.3.22
idna==2.8
ipython==7.4.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10
jsonpath==0.81
MarkupSafe==1.1.0
more-itertools==7.0.0
openapi-codec==1.3.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.9.0
post==2019.3.22
prompt-toolkit==2.0.9
ptyprocess==0.6.0
public==2019.3.22
py==1.8.0
Pygments==2.3.0
PyJWT==1.7.1
pytest==4.4.0
pytest-django==3.4.8
pytz==2018.7
query-string==2019.3.22
request==2019.3.22
requests==2.21.0
rollbar==0.14.7
simplejson==3.16.0
six==1.12.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
appnope==0.1.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.0
Django==2.1.4
django-filter==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.2.0
get==2019.3.22
idna==2.8
ipython==7.4.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10
jsonpath==0.81
MarkupSafe==1.1.0
more-itertools==7.0.0
openapi-codec==1.3.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.9.0
post==2019.3.22
prompt-toolkit==2.0.9
ptyprocess==0.6.0
public==2019.3.22
py==1.8.0
Pygments==2.3.0
PyJWT==1.7.1
pytest==4.4.0
pytest-django==3.4.8
pytest-mock==1.10.4
pytz==2018.7
query-string==2019.3.22
request==2019.3.22
requests==2.21.0
rollbar==0.14.7
simplejson==3.16.0
six==1.12.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
